+++ metaindex/configuration.py
@@ -83,25 +83,43 @@ except ImportError:
 
 ADDONSPATH = DATAPATH / "addons"
 
-CONF_DEFAULTS = {'General': {
-                    'cache': str(CACHEPATH),
-                    'recursive-extra-metadata': "yes",
-                    'collection-metadata': ".metadata, metadata",
-                    'ignore-directories': "\n".join(IGNORE_DIRS),
-                    'ignore-files': "\n".join(IGNORE_FILES),
-                    'accept-files': '',
-                    'index-unknown': 'yes',
-                    'ignore-indexers': '',
-                    'ignore-tags': "Exif.Image.StripByteCounts, Exif.Image.StripOffsets",
-                    'ocr': 'no',
+SECTION_GENERAL = 'General'
+SECTION_SYNONYMS = 'Synonyms'
+SECTION_INCLUDE = 'Include'
+CONFIG_CACHE = 'cache'
+CONFIG_RECURSIVE_EXTRA_METADATA = 'recursive-extra-metadata'
+CONFIG_COLLECTION_METADATA = 'collection-metadata'
+CONFIG_IGNORE_DIRS = 'ignore-directories'
+CONFIG_IGNORE_FILES = 'ignore-files'
+CONFIG_ACCEPT_FILES = 'accept-files'
+CONFIG_INDEX_UNKNOWN = 'index-unknown'
+CONFIG_IGNORE_INDEXERS = 'ignore-indexers'
+CONFIG_IGNORE_TAGS = 'ignore-tags'
+CONFIG_PREFERRED_SIDECAR_FORMAT = 'preferred-sidecar-format'
+CONFIG_OCR = 'ocr'
+CONFIG_MIMETYPES = 'mimetypes'
+
+CONF_DEFAULTS = {SECTION_GENERAL: {
+                    CONFIG_CACHE: str(CACHEPATH),
+                    CONFIG_RECURSIVE_EXTRA_METADATA: "yes",
+                    CONFIG_COLLECTION_METADATA: ".metadata, metadata",
+                    CONFIG_IGNORE_DIRS: "\n".join(IGNORE_DIRS),
+                    CONFIG_IGNORE_FILES: "\n".join(IGNORE_FILES),
+                    CONFIG_ACCEPT_FILES: '',
+                    CONFIG_INDEX_UNKNOWN: 'yes',
+                    CONFIG_IGNORE_INDEXERS: '',
+                    CONFIG_IGNORE_TAGS: "Exif.Image.StripByteCounts, Exif.Image.StripOffsets",
+                    CONFIG_PREFERRED_SIDECAR_FORMAT: '.json, .opf',
+                    CONFIG_OCR: 'no',
                  },
-                 'Synonyms': SYNONYMS,
-                 'Include': {
+                 SECTION_SYNONYMS: SYNONYMS,
+                 SECTION_INCLUDE: {
                  },
                 }
 
 
 class BaseConfiguration:
+    """Convenience wrapper for configparser"""
     TRUE = ['y', 'yes', '1', 'true', 'on']
     FALSE = ['n', 'no', '0', 'false', 'off']
 
@@ -153,17 +171,23 @@ class BaseConfiguration:
 
 
 class Configuration(BaseConfiguration):
+    """Wrapper for BaseConfiguration (aka configparser) with additional convenience accessors"""
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._collection_metadata = None
 
     @property
     def collection_metadata(self):
+        """Return a list of all valid filenames for collection metadata.
+
+        The list is ordered by user's preference for sidecar file format and
+        name of collection metadata file.
+        """
         if self._collection_metadata is None:
-            lst = self.list('General', 'collection-metadata', None)
+            lst = self.list(SECTION_GENERAL, CONFIG_COLLECTION_METADATA, None)
             if lst is not None:
-                lst = set(sum([[fn + store.SUFFIX for store in stores.STORES]
-                               for fn in lst], start=[]))
+                lst = sum([[fn + store.SUFFIX for store in self.get_preferred_sidecar_stores()]
+                           for fn in lst], start=[])
             self._collection_metadata = lst
         return self._collection_metadata
 
@@ -174,49 +198,93 @@ class Configuration(BaseConfiguration):
         E.g. might contain the entry 'title', mapping to ['opf.title', 'id3.title']
         """
         if self._synonyms is None:
-            self._synonyms = dict([(name, self.list('Synonyms', name)) for name in self['Synonyms']])
+            self._synonyms = {name: self.list(SECTION_SYNONYMS, name)
+                              for name in self[SECTION_SYNONYMS]}
         return self._synonyms
 
     def is_sidecar_file(self, path):
         """Return True if the file at path is a sidecar file (either collection metadata or not)"""
-        if not path.is_file:
+        if not path.is_file():
+            # not file? can't be a metadata file
             return False
         if path.name in self.collection_metadata:
+            # the filename *is* the name of a valid collection metadata file
             return True
-        for store in stores.STORES:
-            if path.suffix == store.SUFFIX and len([fn for fn in path.parent.iterdir() if fn.is_file() and fn != path and fn.stem == path.stem]) > 0:
-                return True
-        return False
+        # check whether there are any sidecar files for any of the valid stores
+        return any(path.suffix == store.SUFFIX and
+                    any(fn.is_file() and fn != path and fn.stem == path.stem
+                        for fn in path.parent.iterdir())
+                   for store in stores.STORES)
+
+    def get_preferred_sidecar_stores(self):
+        """Returns a list of all available metadata stores, sorted by preference of the user"""
+        preferred = self.list(SECTION_GENERAL, CONFIG_PREFERRED_SIDECAR_FORMAT, '.json')
+        preferred = [stores.BY_SUFFIX[suffix] for suffix in preferred if suffix in stores.BY_SUFFIX]
+        return preferred \
+             + [store for store in stores.STORES if store not in preferred]
 
     def resolve_sidecar_for(self, path):
-        """Return the path of the direct sidecar file for the given path
+        """Get a sidecar file path for this file
 
-        The returned path may point to a file that does not exist yet!
+        Returns a tuple (path, is_collection, store) with the pathlib.Path to the sidecar file
+        (which may or may not exist), a boolean whether or not the sidecar file is a collection
+        file, and the store module that can be used to read/write the metadata to this sidecar.
+
+        May return (None, False, None) in case there is no usable storage. In that case it will
+        write a message into the error log, too.
         """
-        if not isinstance(path, pathlib.Path):
-            path = pathlib.Path(path).expanduser().resolve()
+        # find what type of metadata file should be used
+        sidecar = None
+        is_collection = None
+        all_stores = [(store, hasattr(store, 'store'))
+                      for store in self.get_preferred_sidecar_stores()]
+        usable_stores = [store for store, usable in all_stores if usable]
+        if len(usable_stores) == 0:
+            return None, False, None
 
-        if path.is_dir():
-            metafilenames = self.list('collection-metadata')
-            for metafilename in metafilenames:
-                metafilepath = path / metafilename
-                if metafilepath.exists():
-                    return metafilepath
-            if len(metafilenames) > 0:
-                return path / metafilenames[0]
-            return None
+        prefer_collection = False
 
-        metafilenames = self.list('')
+        # find existing sidecar file
+        for store, usable in all_stores:
+            location = path.parent / (path.stem + store.SUFFIX)
+            if location.is_file() and usable:
+                sidecar = location
+                break
 
-        return path.parent / metafilename
+        # if there was none, find existing collection sidecar file
+        if sidecar is None:
+            for store, is_usable in all_stores:
+                for collection_name in self.collection_metadata:
+                    location = path.parent / (collection_name + store.SUFFIX)
+                    if location.is_file():
+                        if is_usable:
+                            sidecar = location
+                            is_collection = True
+                            break
+                        prefer_collection = True
+                if sidecar is not None:
+                    break
+        # still none? just take the first preferred sidecar store and create a sidecar file
+        if sidecar is None:
+            if prefer_collection:
+                sidecar_name = self.collection_metadata[0]
+                is_collection = True
+            else:
+                sidecar_name = path.stem
+            sidecar = path.parent / (sidecar_name + usable_stores[0].SUFFIX)
+
+        return sidecar, is_collection, stores.BY_SUFFIX[sidecar.suffix]
 
     def load_mimetypes(self):
-        extra_mimetypes = [pathlib.Path(fn.strip()).expanduser().resolve()
-                           for fn in self.list('General', 'mimetypes', "", "\n")]
+        """Load the user-configured extra mimetypes"""
+        extra_mimetypes = [str(pathlib.Path(fn.strip()).expanduser().resolve())
+                           for fn in self.list(SECTION_GENERAL, CONFIG_MIMETYPES, "", "\n")]
         if len(extra_mimetypes) > 0:
-            mimetypes.init(extra_mimetypes)
+            mimetypes.init(files=extra_mimetypes)
 
-    def load_addons(self):
+    @staticmethod
+    def load_addons():
+        """Load the indexer addons"""
         # load indexer addons
         if ADDONSPATH.exists():
             prev_sys_path = sys.path.copy()
@@ -227,10 +295,10 @@ class Configuration(BaseConfiguration):
                     importlib.import_module(item.stem)
             sys.path = prev_sys_path
 
-
     def ignore_indexers(self):
+        """Remove all indexers that the user configured to be ignored"""
         from metaindex import indexer
-        indexer.remove_indexers(self.list('General', 'ignore-indexers', ''))
+        indexer.remove_indexers(self.list(SECTION_GENERAL, CONFIG_IGNORE_INDEXERS, ''))
 
 
 def load(conffile=None):
@@ -263,4 +331,3 @@ def load(conffile=None):
     config.ignore_indexers()
 
     return config
-